MSC_Thesis/requirements.txt
absl-py==1.0.0
ale-py==0.7.5
anyio==4.2.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
azure==1.0.3
azure-common==1.1.28
azure-mgmt==0.20.2
azure-mgmt-common==0.20.0
azure-mgmt-compute==0.20.1
azure-mgmt-network==0.20.1
azure-mgmt-nspkg==3.0.2
azure-mgmt-resource==0.20.1
azure-mgmt-storage==0.20.0
azure-nspkg==3.0.2
azure-servicebus==0.20.1
azure-servicemanagement-legacy==0.20.2
azure-storage==0.20.3
Babel==2.14.0
backcall==0.2.0
baseline==1.2.1
baselines==0.1.6
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==1.2.2
colorama==0.4.5
comm==0.2.1
contourpy==1.1.1
cycler==0.12.1
Cython==0.29.37
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.9
distlib==0.3.9
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
fasteners==0.19
fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==23.1.21
fonttools==4.47.2
fqdn==1.5.1
fsspec==2023.12.2
future==1.0.0
gast==0.3.3
glfw==2.8.0
google-auth==2.27.0
google-auth-oauthlib==0.5.0
google-pasta==0.2.0
grpcio==1.48.2
gym==0.21.0
gym-notices==0.0.8
gym-retro==0.8.0
gym-super-mario-bros==7.3.0
gymgrid2==1.2.6
h5py==3.1.0
idna==3.6
imageio==2.35.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.12.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.12
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.13.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
lockfile==0.12.2
Markdown==3.5.2
MarkupSafe==2.1.4
matplotlib==3.4.3
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
mujoco-py==2.1.2.14
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
neat-python==0.92
nes_py==8.2.1
nest-asyncio==1.6.0
networkx==3.1
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.22.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
overrides==7.7.0
packaging==23.2
pandas==1.3.3
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==24.0
pip-install==1.3.5
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
progressbar2==4.5.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyglet==1.5.21
Pygments==2.17.2
PyOpenGL==3.1.7
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-utils==3.8.2
pytz==2023.4
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
rsa==4.9
scikit-build==0.17.6
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.11.2
Send2Trash==1.8.2
setuptools==65.5.0
six==1.15.0
sniffio==1.3.0
soupsieve==2.5
stable-baselines3==1.6.0
stack-data==0.6.3
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
terminado==0.18.0
threadpoolctl==3.4.0
tinycss2==1.2.1
tomli==2.0.1
torch==2.1.2
torch-tb-profiler==0.4.3
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
triton==2.1.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.3.0
tzdata==2023.4
uri-template==1.3.0
urllib3==2.2.0
virtualenv==20.28.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wheel==0.42.0
wrapt==1.12.1
zipp==3.17.0
zmq==0.0.0
